New version of LAJ.
From Cathy: I have posted an updated version of Laj, which exits with an error message if there is an integer overflow in building the
composite y-axis for the dotplot (overflows for individual sequences would already be caught by the parser).  I ran it on your file, and
indeed that was the problem.


Former-commit-id: 9c53d9f33a264edb14eb5ae868ba381192d8365f