Update LAJ to latest version.
As per Cathy:
This one deals with the integer overflow by presenting a dialog box for filtering the input data.

It was a little tricky trying to make the dialog's behavior as intuitive as possible under
different usage scenarios; it's still not ideal but hopefully it won't be too confusing.

Please give it a try and let me know if you find any problems.


Former-commit-id: 6f38ea9eef1cd0d8ab663fe7df5befbecc3dd50a