Adds a new config option, pbs_stage_path, that allows you to stage files to/from a directory other than the database/files/ (or whatever you have it configured as) directory.  It's prep work for creating a hierarchical structure for datasets.
This is necessary because there is no other way to force PBS stage
in/out to create missing subdirectories.  So now we stage to somewhere
else (like /tmp), create the dirs in the pbs script, and symlink back
to the stage location.  We could move the dataset if necessary, but a
symlink will be faster if the stage location is on a different
filesystem.


Former-commit-id: 6dd2f573ab75c6a9ddea941af11c5f5df6f0dac0