A DATABASE UPDATE IS REQUIRED:
ALTER TABLE job ADD COLUMN pbs_job_id VARCHAR(255);

Galaxy will now find jobs lost when the server stopped and attempt to
finish them.  This works regardless of whether or not you track jobs in
the database.

A few minor things:

Killed a bug whereby a large number of jobs would be run if
track_jobs_in_database was on and a job was in the NEW state for more
than one loop of the queue monitor (a process was executed for every
loop of the monitor).

command_line is saved to the database when the job is prepared, instead
of when it finishes.

The PBS script, stdout and stderr files are no longer a random name,
they are named with the Galaxy job id.


Former-commit-id: 4724c1cd4d96abdc194c07a46e9a246134361f7e