Updates for Python 2.5 compatibility.
phastOdds was using numarray, test passes with numpy.  Not really a
requirement for py2.5, but it was the only tool not using numpy.

cElementTree is part of the standard distribution in py2.5, updated
megablast xml parser accordingly.

Need simplejson > 1.4 for py2.5 - all tests pass w/ 1.5.

The only showstopper now is Cheetah...


Former-commit-id: 173da86d539b21165e6f6a8f43dcc51a4797fbda