Stop (kill) local processes when the parent output dataset is deleted.
Implemented through a "JobStopQueue," a simplified copy of the
JobQueue.  Done as a queue so dataset deletes don't block while waiting
for the job to terminate.

Changes the PBS stop method to an event instead of a poll, should
reduce database activity.  Prevents jobs in the "new queue" from
running if their input or output datasets are deleted.

Adds a new 'deleted' state to both Job and Dataset objects.  Reports
may need to be modified.


Former-commit-id: ccc050d361ef8a5c746b01f7fc69f4437f8acf5c