Integrate the egg library with Galaxy.
pkg_resources.require is overridden by eggs.require.  You can now do
the following to use any egg in eggs.ini:

from galaxy import eggs
import pkg_resources
pkg_resources.require( 'foo' )

If you don't have it and it's downloadable from eggs.g2.bx.psu.edu,
it'll be downloaded.  If not, it'll raise an EggNotFetchable.  If it's
not in eggs.ini it'll either succeed or raise a DistributionNotFound as
usual.

This also silently ensures that if you require( 'Cheetah' ), ONLY the
version that is listed in eggs.ini will be used.  This should alleviate
problems that have arisen from people's site-installed differing versions.

This won't actually work until I commit the rest of my changes on this
stuff.


Former-commit-id: 9ecca6f00ba9644dfb5a19c2eab0ccccfc7a77b6