Fix Errors reported by Cathy regarding the GMAJ tool.
With these notes:

User's can now specify the warnings that they which to report ('nowarns').  There are likely more options but they are unknown.

Using Firefox 1.5.0.12 on Ubuntu with no extensions/addons installed gives me no issues with loading the applet while Javascript is enabled or disabled (works for me).


Former-commit-id: e128976bf298bc54c3b2cf784343aa4f02ee7a61