Allow for multiple optional outputs to be specified in a tool's xml file by specifying 'filter' tags.  The text contents of the filter tag are evaled. If the result is False, the output will not be created. A NoneDataset object is provided in the param_dict for ease of constructing command lines.
This allows a variable number of output files to be created before the tool is run in cases when the number of outputs is static and can be determined from tool parameters.

Example of an output with filter:

<data format="txt" name="optional_output">
    <filter>some_parameter_name == 'some_parameter_value'</filter>
</data>

The output dataset, optional_output, will only be created when the tool parameter 'some_parameter_name' is 'some_parameter_value'.


Former-commit-id: 47f0b5579e4409c2fef0760167142cc0b54f9099