More secure 'display at' functionality - allows "private" datasets to be displayed at external sites without making them public.  Here's how:
1. Replace the 'display at UCSC <site>' direct link with one back to Galaxy at /dataset/<id>/display_at/ucsc_<site>
2. When the client requests the above path, if the dataset is "private", add a record to the new history_dataset_association_display_at_authorization table.
   note: DATASET_MANAGE_PERMISSIONS is required for this step.  If the user doesn't have appropriate permissions, an error is displayed.
   note 2: "public" datasets don't get a record since they are already readable by external sites.
3. Send the client a redirect to the appropriate external URL.
4. When the external resource requests the dataset (via /root/display_as), allow access if:
   a. an associated (by hda id) record exists in the above table AND
   b. the remote host matches one of the known valid servers for that <site> AND
   c. it's been no longer than 60 seconds since the record was updated (e.g. the lame host authorization expires after 60 seconds)

During the 60 second window, anyone could potentially view the authorized
dataset via UCSC, but they'd have to know the ID *and* know that it was just
authorized.  In the future, it'd be more secure to exchange keys with external
sites or similar strategies, but this would require modifications to these
external applications.

This new method is implemented for UCSC, other display apps will need to be
handled as well, but this should be as simple as adding those sites/servers to
the list in lib/galaxy/security/__init__.py and modifying the link generation
in the display app.


Former-commit-id: a819084bd2d3d2c2f0a68289c7126378f6328b88