First pass at allowing explicit datatype conversion to be specified.  This can be used for e.g. providing a tool with a non-metadata based index file.
Defined as part of a DataToolParameter:

        <param name="input1" type="interval" label="An Interval File">
            <converter name='input1_as_a_bed_file' type='bed'/>
        </param>

Both the original input as well as the converted dataset can be accessed like:

        <command>some_binary $input1 $input1_as_a_bed_file </command>

if $input1 is already BED, it will be used for input1_as_a_bed_file.

The name is be placed in the dictionary space of the data input parameter's parent; so for Grouping objects, e.g. a repeat:


        <repeat name="queries" title="Query">
            <param name="input2" type="data" label="Select" >
            <converter name='input2_as_a_bed_file' type='bed'/>
        </param>
        </repeat>

is accessed like (putting both the original and converted dataset as arguments on the command line):
  <command> ...
        #for $q in $queries
            ${q.input2} ${q.input2_as_a_bed_file}
        #end for
  ... </command>



See notes in code in commit for additional comments.


Former-commit-id: 502b800046c87f37748393d9412d35ce1568bebd