Add the raw uncompressed .js files needed for rgGRR svg display - the packed ones were already there
Fix rgManQQ so the manhattan plot is not called for data without chromosome and offset

Use LD reduced data for IBD/GRR for faster performance and better resolution of close relationships
Add an LD pruning and thinning step to rgGRR - the target composite pbed or lped has the thinned files permanently added for reuse.
This is skipped for small numbers of markers since the plink --thin 0.1 option applied to tinywga.pbed will return an empty file
which is not really ideal.


Former-commit-id: 3e5b0bd2894af4a3f023ad112aac93dd935d9b0d