Fixes for performance problems with display link generation.
1) In "ucsc_links", "generate_display_viewport" was being called before the
check for configured UCSC browsers matching the dataset's build. Thus, Galaxy
would scan the start of the file for every dataset, regardless of whether UCSC
links were displayed in the UI

2) When generating the viewport, if an exception was thrown parsing the
chrom/start/end columns (for example, if the metadata was set wrong causing a
non-numeric string to be found in the start column) the loop would continue
WITHOUT executing the scan to end of line and update of the number of lines
processed. Thus, on any file with incorrect metadata, every line of the file
would be read and split.


Former-commit-id: fbcefb81a4f82eeb30b3ea2b1dcb5806caa65488