Use a job to export a history to an archive file. Exporting a history is now enabled in the history options panel, but importing a history is not yet a job and remains disabled. When a user exports a history, a link is provided that can be used to retrieve the archive or, in the future, moved to another Galaxy instance.
The job that exports a history creates (a) a dataset for storing the archive and (b) multiple temporary attribute files to store the necessary data for exporting a history's datasets, jobs, etc. The temporary files are cleaned up when the job ends. Currently, there is no cleanup for history archive files.


Former-commit-id: 7b0e7f4adde87f1e9a8cf0da624d44b2363728d1