No longer allow Datasets to have DatasetPermissions set such that no roles are associated with the DATASET_MANAGE_PERMISSION.
Include the automatic creation of a new DatasetPermission for an hda where the DATASET_MANAGE_PERMISSION permission is associated with the hda.history.user's private role if the hda has no roles associated with the DATASET_MANAGE_PERMISSION permission.  The creation of the DatasetPermission occurs when the user clicks the pencil icon for the hda.


Former-commit-id: 7498d6f031429183c417ad248313321539455ce0