Updates for picard tools 1.56 - they changed the params for CollectInsertSizeMetrics.jar - bless them - so you will need to update your local picard install in tool-data/shared/jars/picard
Fixes to picard_wrapper.py so if a picard tool returns a non zero exit code, it will raise an error state.
Sadly, this doesn't help the CollectInsertSizeMetrics problem it was meant to address - it will die
with no outputs if given an input containing unpaired reads - but the exit code is still zero.
Talk about irony.
At least it should work right if they ever get around to setting exit codes to indicate a problem.


Former-commit-id: ef88a35e9b2b97cc4b8c6180deb0803c091633b6