Patch dynamic_options.py to just read just the first megabyte of an input dataset (if it is larger than that). Otherwise Galaxy will attempt the read the entire file into memory, this can causes crashes and freezeups without the user even selecting a bad input (for instance, if a large FASTQ file is the top most history item and a txt input is sought, Galaxy will freezeup as soon as the tool is clicked because it will try to load the options for the top item immediately by default).
Though the implementation is not as clean, I am restricting it to a
megabyte instead of a number of lines/options in case the top most history
item has no line breaks. I imagine in that case the same problem could
occur even if the input is restricted to a reasonable number of lines
(say 100).

--HG--
branch : dynamic_options_memory_fix


Former-commit-id: f5f519f40702994a708999989f5db2b7609fa383