Implement a mechanism to secure access to lwr. If enabled on the server, a private token must be passed along with all interactions. This token may be specified as part of the job runner url as such:
wintool = lwr://https://token@host:8913/

This token will not protect the client against a rouge server that has
exploited say a DNS spoofing vulnerability. It is only meant as a
mechanism for the lwr to ensure the client is authorized. This feature
should be used in conjunction with SSL to prevent a "man-in-middle"
from intercepting this token.


Former-commit-id: 9a09a17d9b4cb3d9e091345dfad531aab0091f1e