Adding the initial componenents of an extended metadata system for datasets.
The idea behind an 'extended metadata' for datasets is to provide a mechanism to encode
and index user generated metadata and keep it attached to files. Users can provide a
JSON style data structure which will then be indexed and stored. Later this index can
be used to search and discover datasets via a pseudo-formal schema.

This patch introduces the tables, and makes it possible to inject user metadata into a file
load using the 'extended_metadata' field in the query.


Former-commit-id: de5057aa87073c664555fe68d327bcc8ab0d9863