Add the ability to import a repository capsule that was exported from a tool shed into another tool shed.  Upon initial upload, the capsule is inspected and validated to ensure it is a valid exported capule.
Also, since exporting a capsule allows for including repository dependencies, the Tool Shed into which the capsule is being imported will be inspected to see if any of the repositories defined by any of the capsule's exported repository archives already exists in the Tool Shed.  Any repository that already exists will not be "overwritten" or altered in any way, but the existing repository will be used.

Since repositories that were exported into the capsule are associated with a user (the owner), the user importing the capsule must have the authorization to create the repository in the current Tool Shed with that same owner.  If the current user is an admin user or a member of the IUC, all repositories will be created no matter the owner.  Otherwise, only repositories whose associated owner is the current user will be created.

Repositories whose import process results in an error will not typically halt the import process for
 any remaining repository archives in the capsule.  However, errors encoutnered with importing a rep
ository that is a dependecy of a later repository may force manual fixes in some cases.  For example
, if a capsule includes 2 repsitories that are defined as circular dependencies, one of the repisoto
ry's repository_dependency.xml file will need to be manually altered to add the appropriate changese
t revision since it was not defined at the time the repository was imported.

The Tool Shed API will be enhanced to accommodate this new feature shortly.


Former-commit-id: cb90eb11cd7f8cc6303b3e05fc1b62557353dcf2