Fix small bug in handling of 'brand' text.
This was changed in 6d53f5067dd6d1283895fedaeca2161d1b99583b, after that commit the brand was no longer preceeded by a / in the masthead the way it is in the title. I assume this was an oversite and not intentional, if intentional feel free to backout of this commit.


Former-commit-id: 4eea913409b5f5f898bcd09982fe5179246c1513