Bring workflow parameter context inline with tool.
When tool parameters are passed other_values when rendering tool forms - other_values is the context of other tools at their position in the tree, when rendering workflows other_values is the global state of the tool. The result is different handling of say parameter value filtering for tools and workflows. This changeset switches workflows/run.mako to use the same recursive calculation of context/other_values as tool_form.mako uses.


Former-commit-id: 42731ea66eba78ed9660e13f9b11141615dafa5d