Drop unit incorrect unit tests added in 410a13e and 01933aa.
Workflow run template was causing these methods to be called wrong (and has been for a LONG time), this in turn caused me to misunderstand what the spec for how other_values in tool parameters should operate when writing the tests.

The real underlying bug should be fixed with 55cf8bb.


Former-commit-id: b923a4af8eae0a98388c47318012573ad26da421