Improved (different at least) extra_files_path handling for dataset outputs. $output.extra_files_path will now mimic $output.files_path behavior - writing these files to the working directory to have them copied over later. This is more correct when using outputs_to_working_directory, running jobs as the real user, running jobs through LWR/Pulsar, or when using the NestedObjectStore. This is less efficient the 90% of the time when none of these conditions hold.
There is still an open issue that for inputs -  $input.extra_files_path returns a non-exist path for DiskObjectStores but throws an error for NestedObjectStores. This behavior could be synchronized by this patch (https://gist.github.com/jmchilton/d05bcf02874762cdb829) - though I do not know if this is the intended behavior of NestedObjectStores.


Former-commit-id: 074b6203fd3bef9c7c94c49e81a914143d351a5f