Bugfix for 0750859: Don't generate duplicate workflow slugs for private workflows. See full conversation on https://bitbucket.org/galaxy/galaxy-central/commits/313c1b944d18ceb3988d06fe6483b272.
Would like to better and allow slugs of deleted items to be reused - they cannot be for workflows for instance. Because the slugs are not really handled uniformly though - pages do not have this problem and deleted pages can have their slugs reused because this is included in a check in the web controller which directly assigns valid slugs at that point. Slugs for deleted histories cannot be re-used - but they couldn't be prior to this commit - so this isn't making things worse in that respect.

--HG--
branch : stable


Former-commit-id: d77e5b2670cd72e20e2ba376a075a166fa03b772