Passed port and set URL path for ipython To allow for server administrators to easily secure IPyNBs, we have to stop serving over ports, and start serving under a sub-path. This change will require the introduction of apache configuration as a hard requirement for rewriting URLs properly to the backend.
For a given Port $P, IPyNBs are started, listening on 127.0.0.1:$P/ipython/$P/. These are accessed
through that url within galaxy. For the apache style configuration, we remove the first $P and
simply access notebooks at "127.0.0.1/ipython/$P/" which allows the administrator to apply SSL to
the /ipython/ path, thereby securing the notebooks.


Former-commit-id: cdf156267c8005bb6e5112e08c3cd4e519c51bb7