Possible fix for CORS/cookie issue. I noticed that the cookie was indeed being set by the POST request. However, that cookie was essentially ignored in the subsequent requests.
I looked for places to interrupt it, could I just stop after the POST and not execute any redirects.
That did not appear to be an option due to HTTP/XHR spec. After a bit of searching I came up with
this link: http://stackoverflow.com/a/14462615 which seems to solve the issue. Mentioned here:
https://api.jquery.com/jQuery.ajax/

Former-commit-id: 7577a00267806fa232fc17b7d25669595d28eb6b