Allow for shared secret between Galaxy and upstream proxies
This prevents an impersonation attack available on Galaxies with command line
users and REMOTE_USER authentication. It is not uncommon to run Galaxy on a
"lab server" (or similar), where there are command line users in addition to
the Galaxy service. With REMOTE_USER authentication, Galaxy blindly (and
correctly) trusts the upstream value of REMOTE_USER, enabling such attacks as:

curl -H "REMOTE_USER: admin.user@domain.edu" http://localhost:8000/galaxy/

This changest adds a shared secret between Galaxy and the upstream proxy,
thereby preventing that attack. Only requests from the upstream proxy with the
correct secret key are recognised, everything else is tossed out with a 403
Forbidden.

This was implemented as part of the REMOTE_USER stack as the situation only
applies if that middleware is in use.


Former-commit-id: 7d9b62fdd7bac615641150362053899a5035697e